Extract API key tables, sheets and scope picker into their own components
The API keys settings route had grown to 1000 lines holding twelve
components. The tables, the create and edit sheets, the token reveal
dialog and the shared form fields now live in their own folder.

Adds a smoke test for the page, which had no end-to-end coverage.
