+++ packages/dashboard/e2e/api-keys.spec.ts
@@ -1,0 +1,60 @@
+import { expect, test } from '@playwright/test'
+import { gotoIndex, login } from './helpers'
+
+const API_KEYS_PATH = (storeId: string) => `/${storeId}/settings/api-keys`
+const CTA = /new key/i
+
+test.describe('api keys', () => {
+  test('lists the publishable and secret sections', async ({ page }) => {
+    const creds = await login(page)
+    await gotoIndex(page, API_KEYS_PATH(creds.store_id), CTA)
+
+    await expect(page.getByText('Publishable keys', { exact: true })).toBeVisible()
+    await expect(page.getByText('Secret keys', { exact: true })).toBeVisible()
+  })
+
+  // A secret key is the default choice, so the scope picker has to be there.
+  test('creates a secret key and reveals its token once', async ({ page }) => {
+    const creds = await login(page)
+    await gotoIndex(page, API_KEYS_PATH(creds.store_id), CTA)
+
+    await page.getByRole('button', { name: CTA }).click()
+    const sheet = page.getByRole('dialog')
+    await expect(sheet.getByText(/create api key/i)).toBeVisible()
+
+    const name = `CI key ${Date.now()}`
+    await sheet.locator('#api-key-name').fill(name)
+    await expect(sheet.getByText(/full access/i)).toBeVisible()
+    await sheet.getByText(/read all \(read_all\)/i).click()
+
+    await sheet.getByRole('button', { name: /create key/i }).click()
+
+    // The plaintext token is shown exactly once, behind its own dialog.
+    await expect(page.getByText(/copy.*secret|save.*secret/i).first()).toBeVisible({
+      timeout: 15_000,
+    })
+    await page.getByRole('button', { name: /^done$/i }).click()
+
+    await expect(page.getByText(name)).toBeVisible()
+  })
+
+  test('renames a key through the edit sheet', async ({ page }) => {
+    const creds = await login(page)
+    await gotoIndex(page, API_KEYS_PATH(creds.store_id), CTA)
+
+    // Publishable keys are seeded, so there is always a row to rename.
+    const row = page.locator('tbody tr').first()
+    await expect(row).toBeVisible({ timeout: 15_000 })
+    await row.getByRole('button').last().click()
+    await page.getByRole('menuitem', { name: /^edit$/i }).click()
+
+    const sheet = page.getByRole('dialog')
+    await expect(sheet.getByText(/edit api key/i)).toBeVisible()
+
+    const renamed = `Renamed ${Date.now()}`
+    await sheet.locator('#edit-api-key-name').fill(renamed)
+    await sheet.getByRole('button', { name: /^save$/i }).click()
+
+    await expect(page.getByText(renamed)).toBeVisible({ timeout: 15_000 })
+  })
+})
